refactor(cli-runtime): simplify post-0.37.8 surface
Split the largest runtime files into focused modules and trim helper
duplication so each file stays well below the 500-LOC modularity
threshold:

- completion.ts (390 LOC) is split into completion.ts, completionCommandTree.ts,
  completionBash.ts, completionZsh.ts, and completionFish.ts. Public API
  (generateCompletionScript, parseCompletionShell, CLI_COMPLETION_SHELLS) is
  preserved by re-export; snapshot-tested output is byte-identical.
- supabaseAuth.ts (508 LOC) is split into supabaseAuth.ts,
  supabaseAuthBrowser.ts, and supabaseAuthCallbackServer.ts so the OAuth flow,
  loopback HTTP server, and OS-browser launch no longer share a file.
- doctor.ts: collapse the four `Required<Pick<DoctorDependencies, ...>>` per-check
  parameter types into a single ResolvedDoctorDependencies alias and add a small
  failFromError helper to deduplicate the repeated
  `error instanceof Error ? error.message : undefined` pattern.
- history.ts: drop the unused public export of formatCliHistoryLine (only the
  bulk formatter is consumed externally) and tighten the storage walker so the
  readdir error branch lives inside .catch.

All 870 vitest tests pass; no behavior or public re-export changes.
